Split remote
