+++ frictionless/plugins/remote/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import RemotePlugin
+from .control import RemoteControl
+from .loader import RemoteLoader
